+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -24,7 +24,7 @@ import { PetrLunevilloisS12023 } from './20240108_PetrLunevillois';
 import { Pmgf } from './20221102_PMGF';
 import { Pmgf2023 } from './20230502_PMGF';
 import { PmgfLate2023 } from './20231001_PMGF';
-import { SMT2022 } from './SMT2022';
+import { SMT2022 } from './20220414_SMT';
 import { SMT2023 } from './SMT2023';
 import { SMTC2024 } from './SMTC2024';
 import { SMTC2024Passenger } from './SMTC2024Passenger';
